requirements.txt
django
geocoder
djangorestframework